fix(native) add missing function
